pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hierarchical-diffusion-rlhf-caption-refinement"
version = "0.1.0"
description = "A two-stage generative system for image caption generation and refinement using diffusion models and RLHF with hierarchical reward shaping"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Alireza Shojaei"}
]
keywords = [
    "machine-learning",
    "deep-learning",
    "diffusion-models",
    "rlhf",
    "caption-generation",
    "multimodal",
    "reinforcement-learning",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "torch>=2.0.0",
    "transformers>=4.30.0",
    "diffusers>=0.21.0",
    "trl>=0.7.0",
    "accelerate>=0.20.0",
    "datasets>=2.14.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "PyYAML>=6.0",
    "tqdm>=4.65.0",
    "mlflow>=2.5.0",
    "scikit-learn>=1.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/alirezashojaei/hierarchical-diffusion-rlhf-caption-refinement"
Repository = "https://github.com/alirezashojaei/hierarchical-diffusion-rlhf-caption-refinement"

[tool.setuptools]
packages = ["hierarchical_diffusion_rlhf_caption_refinement"]
package-dir = {"" = "src"}

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--strict-markers --cov=hierarchical_diffusion_rlhf_caption_refinement --cov-report=term-missing"

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
